requirements.txt
alembic==1.8.1
anyio==3.6.1
async-timeout==4.0.2
click==8.1.3
colorama==0.4.5
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
flake8==4.0.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
isort==5.10.1
Mako==1.2.1
MarkupSafe==2.1.1
mccabe==0.6.1
packaging==21.3
passlib==1.7.4
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pydantic==1.9.1
pyflakes==2.4.0
PyJWT==2.4.0
pyparsing==3.0.9
python-multipart==0.0.5
redis==4.3.4
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
sqlalchemy2-stubs==0.0.2a24
sqlmodel==0.0.6
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
wrapt==1.14.1
pyensemblrest
ensembl_rest